Split useAIWorkflowApplications test file for maintainability
Break down the 1,047-line useAIWorkflowApplications.test.ts into three focused test files:

- workflow.test.ts (357 lines) - Workflow YAML application tests
- jobCode.test.ts (433 lines) - Job code preview and application tests
- autoApply.test.ts (440 lines) - Auto-application behavior tests

Also update useAIPanelDiffManager test to match new safety check behavior that returns early when job_id is missing.

All 61 tests passing.
